removed another compiler warning
git-svn-id: svn+ssh://svn.openmicroscopy.org.uk/home/svn/ome/trunk@2270 61f9eca8-113b-463c-a1a4-bab9b0a29f14

+++ src/perl2/OME/Image/Pix/libpix.c
@@ -518,7 +518,7 @@ tsize_t scanline;
 	TIFFGetField(tiff, TIFFTAG_IMAGEWIDTH, &tiffWidth);
 	if (tiffHeight != pPix->dy || tiffWidth != pPix->dx) {
 		fprintf (stderr,"Pix->TIFF2Plane:  Tiff file dimensions (%d x %d) don't match OME Image dimensions (%d x %d).\n",
-			tiffWidth,tiffHeight,pPix->dx,pPix->dy);
+			tiffWidth,tiffHeight,(int)pPix->dx,(int)pPix->dy);
 		TIFFClose (tiff);
 		return (0);
 	}
